+++ CHANGELOG.md
@@ -5,7 +5,7 @@ to [Semantic Versioning](http://semver.org/).
 
 
 
-## Unreleased
+## [0.4.0] - 2015-08-23
 
 ### Added